chore: move entrypoint implementation to dedicated file
